Reorganise: test suite's directory structure
