+++ pkg/validators/test_validator_default_channel.go
@@ -1,9 +1,8 @@
-package testutil
+package validators
 
 import (
 	"github.com/mt-sre/addon-metadata-operator/api/v1alpha1"
 	"github.com/mt-sre/addon-metadata-operator/pkg/utils"
-	"github.com/mt-sre/addon-metadata-operator/pkg/validators"
 )
 
 type ValidatorDefaultChannelTestBundle struct{}
@@ -13,7 +12,7 @@ func (val ValidatorDefaultChannelTestBundle) Name() string {
 }
 
 func (val ValidatorDefaultChannelTestBundle) Run(mb utils.MetaBundle) (bool, error) {
-	return validators.ValidateDefaultChannel(&mb)
+	return ValidateDefaultChannel(&mb)
 }
 
 func (val ValidatorDefaultChannelTestBundle) SucceedingCandidates() []utils.MetaBundle {
